add config dir ,  rm userCollection.js
