--- final_project/.gitignore
@@ -1,1 +1,0 @@
-videos_data
