+++ final_project/config/mongoCollections.js
@@ -1,0 +1,25 @@
+const dbConnection = require("./mongoConnection");
+
+/* This will allow you to have one reference to each collection per app */
+/* Feel free to copy and paste this this */
+let getCollectionFn = (collection) => {
+    let _col = undefined;
+
+    return () => {
+        if (!_col) {
+            _col = dbConnection().then(db => {
+                return db.collection(collection);
+            });
+        }
+
+        return _col;
+    }
+}
+
+/* Now, you can list your collections here: */
+module.exports = {
+    reviews: getCollectionFn("reviews"),
+    locations: getCollectionFn("locations"),
+    videos: getCollectionFn("videos"),
+    users: getCollectionFn("users")
+};
